refactors viewhelpers to be available in more places #124
