--- npm-debug.log
@@ -1,39 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/iojs', '/usr/local/bin/npm', 'install' ]
-2 info using npm@3.9.3
-3 info using node@v6.2.1
-4 silly loadCurrentTree Starting
-5 silly install loadCurrentTree
-6 silly install readLocalPackageData
-7 silly rollbackFailedOptional Starting
-8 silly rollbackFailedOptional Finishing
-9 silly runTopLevelLifecycles Starting
-10 silly runTopLevelLifecycles Finishing
-11 silly install printInstalled
-12 verbose stack Error: Failed to parse json
-12 verbose stack Unexpected token '"' at 30:5
-12 verbose stack     "coverage",
-12 verbose stack     ^
-12 verbose stack     at parseError (/usr/local/lib/node_modules/npm/node_modules/read-package-json/read-json.js:390:11)
-12 verbose stack     at parseJson (/usr/local/lib/node_modules/npm/node_modules/read-package-json/read-json.js:79:23)
-12 verbose stack     at /usr/local/lib/node_modules/npm/node_modules/read-package-json/read-json.js:48:5
-12 verbose stack     at tryToString (evalmachine.<anonymous>:449:3)
-12 verbose stack     at FSReqWrap.readFileAfterClose [as oncomplete] (evalmachine.<anonymous>:436:12)
-13 verbose cwd /Users/Daniel/dwyl/ldmw/app
-14 error Darwin 14.5.0
-15 error argv "/usr/local/bin/iojs" "/usr/local/bin/npm" "install"
-16 error node v6.2.1
-17 error npm  v3.9.3
-18 error file /Users/Daniel/dwyl/ldmw/app/package.json
-19 error code EJSONPARSE
-20 error Failed to parse json
-20 error Unexpected token '"' at 30:5
-20 error     "coverage",
-20 error     ^
-21 error File: /Users/Daniel/dwyl/ldmw/app/package.json
-22 error Failed to parse package.json data.
-22 error package.json must be actual JSON, not just JavaScript.
-22 error
-22 error This is not a bug in npm.
-22 error Tell the package author to fix their package.json file. JSON.parse
-23 verbose exit [ 1, true ]
